channel/channel-backend/requirements.txt
# Framework
django==2.2.9
django-environ==0.4.5
django-redis==4.11.0
django-redis-sessions==0.6.2
djangorestframework==3.9.4
rest-condition==1.0.3
argon2-cffi==19.2.0

# PostgreSQL
psycopg2==2.8.2
psycopg2-binary==2.8.2

# Encryption
pycryptodome==3.9.7
numpy==1.16.3

# Celery
celery==4.4.0
flower==0.9.3

# Http request
requests==2.23.0

# WebScoket
channels==2.3.1
channels_redis==2.4.1
